Restructure project for release
