+++ .github/workflows/python-app.yml
@@ -32,10 +32,11 @@ jobs:
 
     - name: Lint with pylint
       run: |
-        pylint src/ tests/ integration/
+        pylint src/til_page_builder/ tests/ integration/
 
     - name: Test with pytest
       run: |
+        export PYTHONPATH=src/:$PYTHONPATH
         pytest
         
     - name: Format code
