+++ pyproject.toml
@@ -12,3 +12,26 @@ extend-exclude = '''
   | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
 )
 '''
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[project]
+name = "til_page_builder"
+version = "1.0.3"
+authors = [
+  { name="Amnish Singh Arora", email="amnishsingh04@gmail.com" },
+]
+description = "A command-line tool for authoring \"Today I Learned\" posts in Markdown, which can be converted to HTML for publishing on the web."
+readme = "README.md"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: MIT License",
+    "Operating System :: OS Independent",
+]
+
+[project.urls]
+Homepage = "https://github.com/pypa/sampleproject"
+Issues = "https://github.com/pypa/sampleproject/issues"
