+++ pytest.ini
@@ -5,7 +5,7 @@ minversion = 7.0
 testpaths = tests integration
 
 # Directories to look for the source code
-pythonpath = src
+pythonpath = src/til_page_builder
 
 [pytest-watch]
 # Re-run after a delay (in milliseconds), allowing for
