+++ src/til_page_builder/version.py
@@ -1,2 +1,2 @@
 __app_name__ = "TIL Page Builder"
-__version__ = "0.1.0"
+__version__ = "1.0.3"
