+++ tests/builder/toc_generator/test_heading_item.py
@@ -1,4 +1,4 @@
-from builder.toc_generator.heading_item import HeadingItem
+from til_page_builder.builder.toc_generator.heading_item import HeadingItem
 
 
 class TestHeadingItem:
